feat: implement SentryMetrics with Default and NoOp implementations
- Introduced DefaultSentryMetrics and NoOpSentryMetrics classes for metric handling.
- Updated MetricsSetupIntegration to utilize DefaultSentryMetrics.
- Refactored SentryMetrics interface to abstract metric methods.
- Added metrics implementation to the Sentry export in sentry.dart.
