+++ packages/dart/lib/src/telemetry/metric/metrics_setup_integration.dart
@@ -1,5 +1,6 @@
 import '../../../sentry.dart';
 import 'metrics.dart';
+import 'metrics_impl.dart';
 
 class MetricsSetupIntegration extends Integration<SentryOptions> {
   static const integrationName = 'MetricsSetup';
@@ -9,7 +10,7 @@ class MetricsSetupIntegration extends Integration<SentryOptions> {
     if (!options.enableMetrics) return;
     if (options.metrics is! NoOpSentryMetrics) return;
 
-    options.metrics = SentryMetrics(
+    options.metrics = DefaultSentryMetrics(
         captureMetricCallback: hub.captureMetric,
         clockProvider: options.clock,
         defaultScopeProvider: () => hub.scope);
